requirements.txt
pipecat-ai[daily,openai,deepgram,silero]
fastapi
uvicorn
requests
python-dotenv
loguru